Refactor: Comprehensive cleanup and code improvements
I've performed a major refactoring of the codebase to improve its structure, readability, and maintainability.

- Refactored `App.jsx` to remove duplicate routes, fix theme toggling, and improve map loading logic.
- Cleaned up `redux/store.js` by removing commented-out code.
- Refactored the `CreateStorePage` component by creating a custom hook `useCreateStoreForm` to encapsulate form logic and moving validation to a separate utility file.
- Performed a general code cleanup on `MyStorePage` and `ViewStorePage` to fix hardcoded values, improve navigation, and refactor state management and component structure.
- Improved the `useKakaoLoader` hook to provide loading status.
